SWIRL SEARCH 1.6 RC1
Changed email to sid@swirl.today, move swirl_load.py up to root dir, remove install.py and run pip install in the two install scripts; relevancy and processor fixes
