+++ install.sh
@@ -1,7 +1,10 @@
 #!/bin/sh
 
-echo "##S#W#I#R#L##1#.#5##############################################################"
+echo "##S#W#I#R#L##1#.#6##############################################################"
 echo ""
+
+echo "Installing dependencies:"
+pip install -r requirements.txt
 
 echo "Downloading spacy en_core_web_lg..."
 python -m spacy download en_core_web_lg
