+++ swirl.py
@@ -15,25 +15,20 @@ import time
 import signal
 from datetime import datetime
 
-import django
-from sys import path
-from os import environ
-from swirl.utils import swirl_setdir
-path.append(swirl_setdir()) # path to settings.py file
-environ.setdefault('DJANGO_SETTINGS_MODULE', 'swirl_server.settings') 
-django.setup()
+# import django
+# from sys import path
+# from os import environ
+# from swirl.utils import swirl_setdir
+# path.append(swirl_setdir()) # path to settings.py file
+# environ.setdefault('DJANGO_SETTINGS_MODULE', 'swirl_server.settings') 
+# django.setup()
 
-from django.conf import settings
+# from django.conf import settings
 
 module_name = 'swirl.py'
 g_debug = False
 
-class bcolors:
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
+from swirl.banner import SWIRL_BANNER, bcolors
 
 # NOTE: ORDER BELOW IS IMPORTANT
 # NOTE: FIRST IN LIST IS FIRST STARTED
@@ -469,7 +464,7 @@ def setup(service_list):
 
 def main(argv):
 
-    print(f"{bcolors.BOLD}{settings.SWIRL_BANNER}{bcolors.ENDC}")
+    print(f"{bcolors.BOLD}{SWIRL_BANNER}{bcolors.ENDC}")
     print()
 
     parser = argparse.ArgumentParser(description="Manage the SWIRL server")
