+++ swirl/banner.py
@@ -1,0 +1,13 @@
+'''
+@author:     Sid Probstein
+@contact:    sid@swirl.today
+'''
+
+class bcolors:
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+
+SWIRL_BANNER = '##S#W#I#R#L##1#.#6##############################################################'
