+++ swirl/connectors/__init__.py
@@ -1,6 +1,6 @@
 '''
 @author:     Sid Probstein
-@contact:    sidprobstein@gmail.com
+@contact:    sid@swirl.today
 '''
 
 from swirl.connectors.elastic import Elastic
