+++ swirl/mixers/__init__.py
@@ -1,6 +1,6 @@
 '''
 @author:     Sid Probstein
-@contact:    sidprobstein@gmail.com
+@contact:    sid@swirl.today
 '''
 
 from swirl.mixers.relevancy import RelevancyMixer
