+++ swirl/mixers/utils.py
@@ -1,6 +1,6 @@
 '''
 @author:     Sid Probstein
-@contact:    sidprobstein@gmail.com
+@contact:    sid@swirl.today
 @version:    SWIRL 1.x
 '''
