+++ swirl/templates/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>SWIRL SEARCH - Federated Search Engine - HOME</title>
   <meta name="description" content="SWIRL SEARCH Index Page">
-  <meta name="author" content="sidprobstein@gmail.com">
+  <meta name="author" content="sid@swirl.today">
   <meta property="og:title" content="SWIRL SEARCH - Federated Search Engine - HOME">
   <meta property="og:type" content="website">
   <meta property="og:url" content="https://localhost:8000/swirl/">
